+++ magda-web-client/src/Components/Header/HeaderNav.js
@@ -1,7 +1,8 @@
 import React, { Component } from "react";
-// import { NavLink } from "react-router-dom";
+import { Link } from "react-router-dom";
 import AccountNavbar from "../Account/AccountNavbar";
 import { config } from "../../config.js";
+import isExternalURL from "is-url-external";
 
 const headerNavigationPlugins = {
     default: function(nav, i) {
@@ -13,9 +14,20 @@ const headerNavigationPlugins = {
         }
         return (
             <li key={i}>
-                <a {...opts}>
-                    <span>{nav.label}</span>
-                </a>
+                {isExternalURL(opts.href) ? (
+                    <a {...opts} title={`Go to ${nav.label}`}>
+                        <span>{nav.label}</span>
+                    </a>
+                ) : (
+                    <Link
+                        to={opts.href}
+                        target={opts.target}
+                        rel={opts.rel}
+                        title={`Go to ${nav.label}`}
+                    >
+                        <span>{nav.label}</span>
+                    </Link>
+                )}
             </li>
         );
     },